# packages/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project Greenroom API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "greenroom_session"
      }
    },
...
  },
  "paths": {
    "/api/session": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current identity and capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "actor": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "persona": {
                          "type": "string",
                          "enum": [
                            "organizer",
                            "reviewer",
                            "speaker",
                            "public"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "persona"
                      ]
                    },
                    "organizations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          }
                        },
                        "required": [
                          "id"
                        ]
                      }
                    },
                    "eventAccess": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "eventId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "role": {
                            "type": "string",
                            "enum": [
                              "organizer",
                              "reviewer",
                              "speaker",
                              "public"
                            ]
                          },
                          "capabilities": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "enum": [
                                "events:read",
                                "events:create",
                                "events:settings:read",
                                "events:settings:update",
                                "review:manage",
                                "review:evaluate"
                              ]
                            }
                          }
                        },
                        "required": [
                          "eventId",
                          "role",
                          "capabilities"
                        ]
                      }
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "events:read",
                          "events:create",
                          "events:settings:read",
                          "events:settings:update",
                          "review:manage",
                          "review:evaluate"
                        ]
                      }
                    }
                  },
                  "required": [
                    "actor",
                    "organizations",
                    "eventAccess",
                    "capabilities"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "VALIDATION_FAILED",
                            "NOT_FOUND",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "VALIDATION_FAILED",
                            "NOT_FOUND",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/demo-session": {
      "post": {
        "description": "Internal demo-only endpoint; unavailable unless DEMO_MODE is explicitly enabled.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "persona": {
                    "type": "string",
                    "enum": [
                      "organizer",
                      "reviewer",
                      "speaker",
                      "public"
                    ]
                  }
                },
                "required": [
                  "persona"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed demo session established",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "persona": {
                      "type": "string",
                      "enum": [
                        "organizer",
                        "reviewer",
                        "speaker",
                        "public"
                      ]
                    }
                  },
                  "required": [
                    "persona"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "VALIDATION_FAILED",
                            "NOT_FOUND",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "VALIDATION_FAILED",
                            "NOT_FOUND",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Standard error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED",
                            "FORBIDDEN",
                            "VALIDATION_FAILED",
                            "NOT_FOUND",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "type": "string"
                        },
                        "fieldErrors": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}